+++ src/Fable.AST/Fable.fs
@@ -834,11 +834,13 @@ type Expr =
 
     | Unresolved of expr: UnresolvedExpr * typ: Type * range: SourceLocation option
     | Extended of expr: ExtendedSet * range: SourceLocation option
+    | Quote of quotedExpr: Expr * isTyped: bool
 
     member this.Type =
         match this with
         | Unresolved(_, t, _) -> t
         | Extended(kind, _) -> kind.Type
+        | Quote _ -> Any
         | Test _ -> Boolean
         | Value(kind, _) -> kind.Type
         | IdentExpr id -> id.Type
@@ -890,6 +892,7 @@ type Expr =
         | Set(_, _, _, _, r)
         | ForLoop(_, _, _, _, _, r)
         | WhileLoop(_, _, r) -> r
+        | Quote(e, _) -> e.Range
 
 // module PrettyPrint =
 //     let rec printType (t: Type) = "T" // TODO